EdManiaApi/pyproject.toml
[tool.poetry]
name = "edmaniaapi"
version = "0.1.0"
description = ""
authors = ["Amos Kipyegon <amoskip@kabarak.ac.ke>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
whitenoise = "^6.3.0"
drf-yasg = "^1.21.4"
python-dotenv = "^0.21.0"
pytest = "^7.2.0"
djoser = "^2.1.0"
dj-database-url = "^1.2.0"
coverage = "^7.0.5"
black = "^22.12.0"
django = "^4.1.5"
python-environ = "^0.4.54"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
